pyproject.toml
[tool.pdm]
[tool.pdm.scripts]
dev = "python manage.py runserver"
runserver = "python manage.py runserver"
createsuperuser = "python manage.py createsuperuser"
migrate = "python manage.py migrate"
makemigrations = "python manage.py makemigrations"
shell = "python manage.py shell"
test = "python manage.py test"
startapp = "python manage.py startapp {args}"
loaddata = "python manage.py loaddata {args}"
dumpdata = "python manage.py dumpdata {args}"
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "joao gabriel", email = "zala4453523@gmail.com"},
]
dependencies = [
    "django>=4.2.1",
    "djangorestframework>=3.14.0",
    "django-cors-headers>=4.0.0",
]
requires-python = ">=3.10"
license = {text = "MIT"}
